session

+++ src/main/java/com/boot/controller/BoardController.java
@@ -22,20 +22,27 @@ import lombok.extern.slf4j.Slf4j;
 @Slf4j
 public class BoardController {
 	
+	static String checkSession(HttpSession session) {
+		MemDTO mDTO = (MemDTO) session.getAttribute("LOGIN_MEMBER");
+		log.info("@# mDTO=>" + mDTO);
+		if(mDTO == null) {
+			return "redirect:login";
+		}
+		return "OK";
+	}
+	
 	@Autowired
 	private BoardService service;
 	
 	@RequestMapping("/list")
 //	public String list(Model model) {
 	public String list(Model model, HttpServletRequest request) {
-		log.info("@# list()");
-		
-		HttpSession session = request.getSession();
-		MemDTO mDTO = (MemDTO) session.getAttribute("LOGIN_MEMBER");
-		
-		if(mDTO == null) {
+		String result = checkSession(request.getSession());
+		if(!result.equals("OK")) {
 			return "redirect:login";
 		}
+		
+		log.info("@# list()");
 		
 		ArrayList<BoardDTO> list = service.list();
 		model.addAttribute("list", list);
@@ -45,11 +52,8 @@ public class BoardController {
 	
 	@RequestMapping("/write")
 	public String write(@RequestParam HashMap<String, String> param, HttpServletRequest request) {
-		
-		HttpSession session = request.getSession();
-		MemDTO mDTO = (MemDTO) session.getAttribute("LOGIN_MEMBER");
-		
-		if(mDTO == null) {
+		String result = checkSession(request.getSession());
+		if(!result.equals("OK")) {
 			return "redirect:login";
 		}
 		
@@ -61,14 +65,24 @@ public class BoardController {
 	}
 	
 	@RequestMapping("/write_view")
-	public String write_view() {
+	public String write_view(HttpServletRequest request) {
+		String result = checkSession(request.getSession());
+		if(!result.equals("OK")) {
+			return "redirect:login";
+		}
+		
 		log.info("@# write_view()");
 		
 		return "write_view";
 	}
 	
 	@RequestMapping("/content_view")
-	public String content_view(@RequestParam HashMap<String, String> param, Model model) {
+	public String content_view(@RequestParam HashMap<String, String> param, Model model, HttpServletRequest request) {
+		String result = checkSession(request.getSession());
+		if(!result.equals("OK")) {
+			return "redirect:login";
+		}
+		
 		log.info("@# content_view()");
 		
 		BoardDTO dto = service.contentView(param);
@@ -78,7 +92,12 @@ public class BoardController {
 	}
 	
 	@RequestMapping("/modify")
-	public String modify(@RequestParam HashMap<String, String> param, Model model) {
+	public String modify(@RequestParam HashMap<String, String> param, HttpServletRequest request) {
+		String result = checkSession(request.getSession());
+		if(!result.equals("OK")) {
+			return "redirect:login";
+		}
+		
 		log.info("@# modify()");
 		service.modify(param);
 		
@@ -86,11 +105,26 @@ public class BoardController {
 	}
 	
 	@RequestMapping("/delete")
-	public String delete(@RequestParam HashMap<String, String> param, Model model) {
-		log.info("@# delete()");
+	public String delete(@RequestParam HashMap<String, String> param, HttpServletRequest request) {
+		String result = checkSession(request.getSession());
+		if(!result.equals("OK")) {
+			return "redirect:login";
+		}
 		
+		log.info("@# delete()");
 		service.delete(param);
 		
 		return "redirect:list";
 	}
+	
+//	@RequestMapping("/logout")
+//	public String logout(HttpServletRequest request) {
+//		HttpSession session = request.getSession();
+//		if(session == null) {
+//			log.info("이미 세션이 만료되었습니다.");
+//		}else {
+//			session.invalidate();
+//		}
+//		return "login";
+//	}
 }
